config refactor stage 3 - nnictl & Experiment (#4330)
