+++ nni/experiment/config/utils/public.py
@@ -23,6 +23,9 @@ def is_missing(value):
     return isinstance(value, type(dataclasses.MISSING))
 
 def canonical_gpu_indices(indices):
+    """
+    If ``indices`` is not None, cast it to list of int.
+    """
     if isinstance(indices, str):
         return [int(idx) for idx in indices.split(',')]
     if isinstance(indices, int):
@@ -38,9 +41,19 @@ def validate_gpu_indices(indices):
         raise ValueError(f'Negative detected in GPU indices {indices}')
 
 def parse_time(value):
+    """
+    If ``value`` is a string, convert it to integral number of seconds.
+    """
     return _parse_unit(value, 's', _time_units)
 
+def parse_memory_size(value):
+    """
+    If ``value`` is a string, convert it to integral number of mega bytes.
+    """
+    return _parse_unit(value, 'mb', _size_units)
+
 _time_units = {'d': 24 * 3600, 'h': 3600, 'm': 60, 's': 1}
+_size_units = {'tb': 1024 ** 4, 'gb': 1024 ** 3, 'mb': 1024 ** 2, 'kb': 1024, 'b': 1}
 
 def _parse_unit(value, target_unit, all_units):
     if not isinstance(value, str):
